+++ lib/runners/processor/go_vet.rb
@@ -2,6 +2,12 @@ module Runners
   class Processor::GoVet < Processor
     include Go
 
+    Schema = StrongJSON.new do
+      let :runner_config, Schema::BaseConfig.base
+    end
+
+    register_config_schema(name: :go_vet, schema: Schema.runner_config)
+
     def self.ci_config_section_name
       # Section name in sideci.yml, Generally it is the name of analyzer tool.
       "go_vet"
@@ -13,6 +19,11 @@ module Runners
 
     def analyzer_name
       "go_vet"
+    end
+
+    def setup
+      add_warning_for_deprecated_linter(alternative: "GolangCi-Lint", deadline: Time.new(2_020, 3, 31))
+      yield
     end
 
     def analyze(changes)