+++ CHANGELOG.md
@@ -12,6 +12,9 @@ All notable changes to this project will be documented in this file.
 - [FxCop] New support [#731](https://github.com/sider/runners/pull/731)
 - Aggregate log entries by `delete_unchanged_files` method [#765](https://github.com/sider/runners/pull/765)
 - [RuboCop] Use `--out=FILE` option [#769](https://github.com/sider/runners/pull/769)
+- [go-vet][Golint][Go Meta Linter] Add deprecated warning [#767](https://github.com/sider/runners/pull/767)
+- [Goodcheck] Update goodcheck requirement from 2.4.5 to 2.5.0 [#775](https://github.com/sider/runners/pull/775)
+- Add Runners::Config, which is responsible for sider.yml [#763](https://github.com/sider/runners/pull/763)
 - [detekt] New support [#749](https://github.com/sider/runners/pull/749)
 
 ## 0.20.0